Fixed typos in 6.1
architecture of an Network OS -> architecture of a Network OS

there are actually a wide assortment of apps -> there is actually a wide assortment of apps


Also, there are differences between Fig. 30 and the text. In the figure, FlowObjective and FlowRule are written (spaceless). However, in the text, there are multiple times that "Flow Objectives" and "Flow Rules" are mentioned (with space)