fixed typos/grammatical errors in 6.2
particular aspect of network state ->  particular aspect of the network state
achieve fault-tolerance -> achieve fault tolerance
The web site also -> The website also
manage the distributed state and to provide easy access -> manage the distributed state and provide easy access
highly available (run on enough -> highly available (runs on enough

a whole, and then scaled across -> a whole and then scaled across 

Note that the set of ONOS instances are distinct from  -> Note that the set of ONOS instances is distinct from 

instances are able to monitor (read) switch state. -> instances are able to monitor (read) switch states.
when a new switch comes on-line -> when a new switch comes online

key/value pairs and the other is -> key/value pairs, and the other is
have an associated map, but instead indirectly -> have an associated map but instead indirectly

to access network state. -> to access the network state.
inter-connected -> interconnected

several things of note about this dependency graph. -> several things to note about this dependency graph.

forward select packets to the -> forward selected packets to the
 topology is fixed, and known ->  topology is fixed and known
(i.e. how to -> (i.e., how to
and to emit packets back into the network. -> and emit packets back into the network.

but for now we note -> but for now, we note

nexthop mapping -> next-hop mapping

for select network traffic -> for selected network traffic
data-plane forwarding behavior -> data plane forwarding behavior
 High-level specifications, call *intents*, ->  High-level specifications, called *intents*,
over the appropriate paths and then -> over the appropriate paths, and then