fixed typos/grammatical errors in 6.3
First, for every service included in a given configuration of ONOS, there is a corresponding API. -> First, there is a corresponding API for every service included in a given configuration of ONOS.

standardized interfaces, defined independent of ONOS, but supported as -> standardized interfaces, defined independent of ONOS but supported as

from OpenFlow, and hence, is pipeline-aware. The second is pipeline-agnostic, and the focus -> from OpenFlow and hence, is pipeline-aware. The second is pipeline-agnostic and the focus

Next objectives indicate what kind -> The Next objectives indicate what kind

a routing table; and finally the Next -> a routing table; and finally, the **Next**

to exactly what combination of tables in the underlying switch are used to implement -> to exactly what combination of tables in the underlying switch is used to implement 

and IP address combination, without regard for -> and IP address combination without regard for

Filtering, Filter -> **Filtering**, **Filter**
Forwarding -> **Forwarding**
Next -> **Next**

will be unified, and in all likelihood -> will be unified and in all likelihood

flow objectives are a data structure, packaged with -> flow objectives are a data structure packaged with

underlying devices is done elsewhere, and not included in -> underlying devices is done elsewhere and not included in