Fixed grammatical errors and typos in Ch. 7
equipped with software -> equipped with the software
programmable pipelines, but is -> programmable pipelines but is
supports a wide-range of L2/L3 features -> supports a wide range of L2/L3 features
access networks, and (b) -> access networks and (b)

that already exists, and do so in -> that already exists and do so in
protocol like PIM), but does->  protocol like PIM) but does
load balance outgoing packets -> load-balance outgoing packets
application knows the topology, and pushes -> application knows the topology and pushes
within each server rack, and L3 connectivity  -> within each server rack and L3 connectivity

match labeled or unlabeled packets, and to push or pop MPLS labels as needed. -> match labeled or unlabeled packets and push or pop MPLS labels as needed.
 each of the switches have an assigned MPLS id ->  each of the switches has an assigned MPLS id
destination address 10.0.2.1 it is by -> destination address 10.0.2.1, it is by

each IP prefix, and where to find -> each IP prefix and where to find
answers based on global information, and then -> answers based on global information and then
``GET``, ``POST`` and ``DELETE`` calls. -> ``GET``, ``POST``, and ``DELETE`` calls.
SD-Fabric inserted it, with full knowledge -> SD-Fabric inserted it with full knowledge
the group source addresses (``srcs``) and the group sink -> the group source addresses (``srcs``), and the group sink
calls similar to the ones shown above -> calls similar to those shown above

:numref:`Figure %s <fig-fabric>` which shows -> :numref:`Figure %s <fig-fabric>`, which shows
and in practice it makes heavy use -> and in practice, it makes heavy use
in more depth, but at -> in more depth but at
it is worth nothing that -> it is worth noting that
higher level  -> higher-level 
on-going -> ongoing

run in switch forwarding pipeline, rather than  ->  run in the switch forwarding pipeline rather than